requirements.txt
Flask==1.1.2
gunicorn==20.1.0
Jinja2==2.11.3
Werkzeug==1.0.1
pandas==1.2.3
numpy==1.20.2
keras==2.4.3
scikit-learn==0.24.1
scipy==1.6.2
flask_login==0.5.0
flask_migrate==2.7.0
flask_wtf==0.14.3
flask_sqlalchemy==2.5.1
email_validator
python-decouple
python-dotenv~=0.16.0
flask-mail==0.9.1
flask-debugtoolbar==0.11.0
flask-script==2.0.6
WTForms~=2.3.3
SQLAlchemy~=1.4.3
alembic~=1.5.8
setuptools~=54.2.0